Update README.md
using a manual system. Criminal record management system is to allow the administrator of any organization to edit and find out the details of a criminal and keep up to date to his/her profile.It'll also facilitate keeping all the records of criminals,such as their first name, last name, city, phone number, committed crime, e.t.c. So all the information about the criminal will be available in a few seconds. Overall it'll make criminal record management an easir job for management.